Move existing plugins into the plugin foler
Moves DSL, Initializer, GenericAccessor into the plugins folder and
namespace.
