+++ lib/smart_properties.rb
@@ -9,7 +9,7 @@ define_method(:SmartProperties, &Class.new(Module) do
   end
 
   def included(base)
-    [SmartProperties::DSL, *plugins].each { |plugin| plugin.attach(base) }
+    [SmartProperties::Plugins::DSL, *plugins].each { |plugin| plugin.attach(base) }
   end
 
   def inspect
@@ -32,7 +32,7 @@ define_method(:SmartProperties, &Class.new(Module) do
   end
 end)
 
-SmartProperties = SmartProperties(-> { SmartProperties::Initializer }, -> { SmartProperties::GenericAccessors })
+SmartProperties = SmartProperties(-> { SmartProperties::Plugins::Initializer }, -> { SmartProperties::Plugins::GenericAccessors })
 
 ##
 # {SmartProperties} can be used to easily build more full-fledged accessors
@@ -59,12 +59,9 @@ SmartProperties = SmartProperties(-> { SmartProperties::Initializer }, -> { Smar
 module SmartProperties; end
 
 require_relative 'smart_properties/plugin'
-require_relative 'smart_properties/dsl'
-require_relative 'smart_properties/generic_accessors'
-require_relative 'smart_properties/initializer'
+require_relative 'smart_properties/plugins'
 require_relative 'smart_properties/property_collection'
 require_relative 'smart_properties/property'
 require_relative 'smart_properties/errors'
 require_relative 'smart_properties/version'
 require_relative 'smart_properties/validations'
-require_relative 'smart_properties/plugins'
