+++ lib/smart_properties/plugins.rb
@@ -3,4 +3,7 @@ module SmartProperties
   end
 end
 
+require_relative 'plugins/dsl'
 require_relative 'plugins/equality'
+require_relative 'plugins/generic_accessors'
+require_relative 'plugins/initializer'
