+++ lib/smart_properties/plugins/generic_accessors.rb
@@ -1,0 +1,17 @@
+module SmartProperties
+  module Plugins
+    GenericAccessors = SmartProperties::Plugin.new(:include) do
+      def [](name)
+        return if name.nil?
+        name = name.to_sym
+        reader = self.class.properties[name].reader
+        public_send(reader) if self.class.properties.key?(name)
+      end
+
+      def []=(name, value)
+        return if name.nil?
+        public_send(:"#{name.to_sym}=", value) if self.class.properties.key?(name)
+      end
+    end
+  end
+end
